refactor: remove unused Recent Conversations section
