+++ src/components/features/conversation-panel/conversation-status-dot.tsx
